Allow a custom redirect_uri to be set
I want to be able to set a custom redirect_uri for development. That is,
in order to test in development I am using localhost:4567 etc. (as it's
an internal server that I cannot approve). Unfortunately, the app I am
developing against is on a (named) local server, so, without being able
to override the redirect_uri I cannot work on it.

This change allows the user to set the `redirect_uri` either at
configuration time, or, as part of the url /auth/:provider?redirect_uri=foo